+++ create-vtj/templates/app/src/main.ts
@@ -3,21 +3,26 @@ import {
   LocalService,
   createModules,
   NodeEnv,
-  autoUpdate
+  autoUpdate,
+  notify,
+  loading,
+  createAdapter,
+  createServiceRequest
 } from '@vtj/web';
 import { createApp } from 'vue';
 import router from './router';
 import App from './App.vue';
-import './style/index.scss';
 import { name, description } from '../package.json';
+import './style/index.scss';
 
 const app = createApp(App);
-const service = new LocalService();
-
+const adapter = createAdapter({ loading, notify });
+const service = new LocalService(createServiceRequest(notify));
 const { provider, onReady } = createProvider({
   nodeEnv: process.env.NODE_ENV as NodeEnv,
   modules: createModules(),
   service,
+  adapter,
   router,
   dependencies: {
     Vue: () => import('vue'),