+++ packages/materials/src/vant/button.ts
@@ -1,7 +1,7 @@
 import type { MaterialDescription } from '@vtj/core';
 
 const Button: MaterialDescription = {
-  name: 'VantButton',
+  name: 'VanButton',
   alias: 'Button',
   label: '按钮',
   categoryId: 'base',
@@ -34,7 +34,7 @@ const Button: MaterialDescription = {
     {
       name: 'icon',
       title: '左侧图标名称或图片链接，等同于 Icon 组件的 name 属性',
-      setters: 'StringSetter'
+      setters: 'VanIconSetter'
     },
     {
       name: 'icon-prefix',