+++ packages/node/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.9.6](https://gitee.com/newgateway/vtj/compare/@vtj/node@0.9.5...@vtj/node@0.9.6) (2025-01-06)
+
+**Note:** Version bump only for package @vtj/node
+
+
+
+
+
 ## [0.9.5](https://gitee.com/newgateway/vtj/compare/@vtj/node@0.9.4...@vtj/node@0.9.5) (2025-01-04)
 
 **Note:** Version bump only for package @vtj/node